Add detailed view for budget plan
The commit introduces a detailed view for each budget plan. The new additions include some DTOs for displaying the transaction details with each item for every day. Alongside this, it enhances existing repositories to retrieve needed data and implements the necessary managers and builders to assemble the detailed view. A new endpoint in the API has been added to retrieve this detailed view. Exception handling is also included if the requested budget plan is not found or the user has no access to it.