Remove CSS for CreateTransactionDetails, improve layout and async calls
The CSS file for CreateTransactionDetails component has been deleted and replaced with CSS in JS within the component itself. This includes the reorganization of inputs into InputGroup for a cleaner layout and a more intuitive design. Async calls in the handleSubmit function have been updated for better performance. Some minor updates on package dependencies.